rtr_mgr: make rtr_mgr_config a private data structure
Previously the rtr_mgr_config definition was publicly available and
parts from it had to be initialized from the user.

The definition moved to rtr_mgr.c and isn't public anymore. This
prevents that library access attributes of the struct rtr_mgr_config
while they are in use from another thread of the library. It was also
confusing for the user that he only had to initialize parts of
the struct rtr_mgr_config.

In consequence the rtr_mgr_init(..) signature was changed. The
rtr_mgr_config parameters was removed instead it now returns
a pointer to an initialized rtr_mgr_config or NULL if an error occured.
rtr_mgr_free(..) deallocates the memory allocated by the rtr_mgr_conf,
not only from the memory of the attributes in the rtr_mgr_conf.