transport: remove const modifier from tr_{ssh,tcp}_config members
Remove the const modifier from tr_tcp_config and tr_ssh_config elements.
The const modifier for the elements made it hard to follow
the common declaration-after-statement style.
It also stands in the way when the variable was created and the
assignment should be done in a different scope, e.g. in the branch of an
if-statement