pfx: replace socket_id member with a pointer to the rtr_socket
Replace the socket_id member in a pfx_record, with a socket member from
type rtr_socket *. The socket member references the rtr_socket that
retrieved the pfx_record.

The previously stored socket_id also contained the memory address of the
rtr_socket but was from type uintptr_t. Converting it to type rtr_socket *
enables library users to access informations about the rtr_socket that
retrieved a particular record.