rtr_mgr: pass the rtr_socket to the status callback function
The rtr_mgr status callback is now called for every status change of an
rtr_socket and the rtr_socket is passed to the callback function.
The library user knows exactly which server connection changed the
status and can eg update the status of alle server connection as soon as
they changed.