rtrclient: fix segfault, access to out of scope tr_config members
A segfault happened in rtrclient shortly after start.
The ssh_config and tcp_config variables were
defined in the block scope of an if-statement. In this block the memory
address of the config structs were passed to the transport init
functions.
When the config were accessed in a later state of the program again, the
scope of the config variables was expired.

(Thanks to David Mandelberg and Michael Mester for reporting and
locating the cause of the segfault)