debian: don't run "make test" during package creation
The "make test" run failed while package creation  because the rtrlib
shared library couldn't be found.
I suppose it's because of the required -DCMAKE_SKIP_RPATH=True cmake
option in the debian rules file.
For now it's sufficient if the test are run manually and not
automatically when a debian package is created.