spki: fix mismatch of return value type
The return value of spki_table_remove_entry() was declared as int, but
the function the stored the value that it returned in an "unsigned int"
variable.
It also assigned negative values to the unsigned int variable and
returned it then as signed int.

This commit also removes the unnecessary initialization of the rtval
variable. It's always set in the function before used.