scripts: supress cppcheck duplicateExpression warning in packets.c
cppcheck was warning:
	[rtrlib/rtr/packets.c:974] -> [rtrlib/rtr/packets.c:974]: (style) Same
	expression on both sides of '||'.
	[rtrlib/rtr/packets.c:976] -> [rtrlib/rtr/packets.c:976]: (style) Same
	expression on both sides of '||'.
The expression checks if the rtval is the same than enum error values,
that are represented internally as the same integer values.
It should explicitly check for the enum values to ensure that the code
still works if the order of the enum changes.

Therefore suppress the cppcheck warning.