rtrlib: fix pointer arithmetics with void * types
The C standard doesn't allow arithmetics with void * types because
it's an incomplete type.
Fix the parts in packets.c by using a typed ptr for the arithmetics.
This fixes the cppcheck warnings:
	arithOperationsOnVoidPointer:rtrlib/rtr/packets.c(737):(portability)
	'ary' is of type 'void * *'. When using void pointers in
	calculations, the behaviour is undefined.
	arithOperationsOnVoidPointer:rtrlib/rtr/packets.c(739):(portability)
	'ary' is of type 'void * *'. When using void pointers in
	calculations, the behaviour is undefined.
	arithOperationsOnVoidPointer:rtrlib/rtr/packets.c(764):(portability)
	'ary' is of type 'void * *'. When using void pointers in
	calculations, the behaviour is undefined.

This commit also adapts the line number for suppressed cppcheck
messages in the file which moved with this change.