lpfst: convert the lpfst to a shortest prefix search tree
The longest prefix search tree failed to deliver the correct validation
result in the following scenario:

he LPFST looks like this:
             10.100.255.0/24-24, ASN C
             /               \
            /                 \
   128.0.0.0/1-24, ASN A     255.0.0.0/24-24, ASN B

We are trying to verify 255.0.0.0/24, ASN A.
The lpfst_lookup() starts searching at the root node, compares bit 0 of
10.100.255.0 (1100) and 255.0.0.0 (1111, they don't match.
Because bit 0 of the searched prefix is 1 it continues the search at the right
child node and never find the matching pfx record at the right child node.

To fix it, convert it to shortest prefix search tree.
Like before the position of the nodes in the tree are determined by the
prefix.
Either
- if the tree contains 0 nodes, the new node is inserted as root node
  otherwise
- lpfst-pfx finds the node that matches the prefix up to mask_len
  and insert the pfx records to the array of the node if such a node
  doesn't exist
- lpfst-pfx returns the node in the tree that matches the prefix up to
mask_len-N bits and executes the insert operation there

If the mask_len of a node that is inserted at node X is shorter than the
mask_len of X, those both are swapped.
If the mask_len is >= mask of node X, the bit <tree-level> of the prefix is
checked. If the bit is 0 the node is inserted as left child otherwise as
right child.

Therefore the max. level the tree can growth is the biggest used
mask_len.