Add trie data structure
The longest prefix search tree failed to deliver the correct validation
result in the following scenario:

             10.100.255.0/24-24, ASN C
             /                     \
            /                       \
      128.0.0.0/1-24, ASN A     255.0.0.0/24-24, ASN B

We are trying to verify 255.0.0.0/24, ASN A.
The lpfst_lookup() starts searching at the root node, compares bit 0 of
10.100.255.0 (1100) and 255.0.0.0 (1111), they don't match.
Because bit 0 of the searched prefix is 1 it continues the search at the right
child node and never find the matching pfx record in the left child node.

This commit adds a new data structure, a trie, that gives correct validation results.
The key difference between the trie and the lpfst is that the trie root is the shortest
(or one of the shortest) prefix and the leafs are the longest prefixes. This is the opposite
of the lpfst where the root is the longest (or one of the longest) prefix and the leafs are
short prefixes.

The trie data structure works with the same structs as the lpfst. The function implementations
are identical with the exception of 'lpfst_lookup(..)'. lpfst_lookup(..) returns the longest matching
prefix, which is also the first match in the lpfst. For the trie, trie_lookup(..) must traverse the trie
until it hits a leaf and then returns the best match it found on the way.

Since for validation purposes it is more efficient to evaluate the first match in the data structure and
only if the AS does not match continue searching, there is now a new function for the trie: trie_lookup_first_match(..)
This function is also used in the pfx_table_validate_r(..) implementation of the trie.

Other functions of the trie are identical to the lpfst, with the exception that
in '_insert' and '_lookup_exact' the prefix mask comparisons are turned around to place