tests: add test to reproduce github issue #99
Adding the PFX records:
10.100.255.0/24-24 ASN 200
255.0.0.0/24-24 ASN 300
128.0.0.0/1-24 ASN 400

and validating the route 255.0.0.0/24 ASN 400 returns an invalid
validation result. It should be valid because the route is covered by
the pfx records 128.0.0.0/1-24 ASN 400.

This commit also introduces the helper function add_ip4_pfx_record() in the
test_file to allow adding IPv4 records to the pfx table without duplicating the
code all the time.

https://github.com/rtrlib/rtrlib/issues/99