pfx: fix possible memory leak in pfx_table_append_elem()
When appending of a new pfx record to the trie resulted in the creation
of the new node and the allocation of the node->data structure failed
the node was leaked and never freed.
Free the node in this case.

The bug was found by Nils Bars and the patch is based on a commit from
Andreas Reuter.