pfx: fix: user is notified about a new pfx record when appending failed
When the appending of a pfx_record to a tree node failed,
pfx_table_notify_clients() was still called to notify the user about a
new added pfx record.
But in fact no record was added because the append operation failed.
Only call pfx_table_notify_clients() when the append operation was
successful.