rtr_mgr: fix memory leak in rtr_mgr_get_spki
rtr_mgr_get_spki passes a pointer to a variable that lives in the stack
frame of itself as pointer to pointer to spki_table_get_all.
spki_table_get_all treats this argument as pointer to an array and may
realloc it at will. But since the pointer points to a value on the stack
of rtr_mgr_get_spki it is lost when rtr_mgr_get_spki returns. Thus
leading to a memory leak and causing the function to not actually
returning the result, rendering it useless.

This change technically breaks the api, but since this function was not
functional anyway it should not be a problem.

coverity id: 1420793