rtr_mgr: Fix memory leak in rtr_mgr_init
Due to missing validation of the groups_len parameter in rtr_mgr_init
memory allocated for the pfx and spki tables could leak. In the unlikely
case that groups_len is zero no pointer to it would exist after the function
returns. Usually they are passed to rtr_init and stored in the
rtr_sockets, but that does not happen when groups_len is null, since
there is no socket to initialize in that case.

Furthermore this commit fixes a bug in the cleanup which caused a freed
pointer to be returned through config_out, in case of an error
situation.

coverity ids: 1420791 1420783