rtr: Add new parameter connection_state_fp_param_group to rtr_socket to be passed to connection_state_fp.
This new parameter is a pointer to the rtr_mgr_group that this rtr_socket belongs to. This is then used in rtr_mgr_cb to get the group of the socket doing the callback. This was necessary because otherwise we'd have iterate over the linked list to find the group of the socket (rtr_mgr_find_group). This can lead to a deadlock when the rtr_mgr_cb callback execution path came from a place where we already iterate over the linked list, such as rtr_mgr_stop which calls rtr_stop