rtr: Add new rtr_socket state RTR_CLOSED. This state is the new initial state of an rtr_socket.
The state RTR_SHUTDOWN still exists and is still used to shutdown a socket that has already entered the state machine loop (rtr_fsm_start). The new state is necessary because of the following scenario: When rtr_stop is called on a rtr_socket that has been started  with rtr_start, but has not started execution of the state machine loop yet, the rtr_stop state change had no effect since RTR_SHUTDOWN was the initial state. With this new state, the socket can distinguish between its initial state and the RTR_SHUTDOWN state. Before the state machine loop begins, we can check whether the socket has been shutdown by rtr_stop. If so, we need not begin the loop. This bug occured when calling rtr_mgr_remove_group and then right after rtr_mgr_stop. Remove group lead to the next best group being started, and the rtr_stop SHUTDOWN state change came to early for this socket to notice.