{tcp,ssh}_transport: duplicate socket configuration on init
Previously the socket configuration was just a shallow copy of the user
provided struct. This caused problems while developing the frr
integration because there is basically no way of freeing the strings in
it later on, unless they were allocated on the stack, which is not
feasible in larger applications.

The custom strdup implementation is necessary because libc strdup calls
malloc.