packets.c: Fix wrong length of error text in error pdu
Due to a missing cast only the lowest byte of the 4 Byte length of error
text field in the error pdu of the rtrlib protocol was set. The upper 3
Bytes were basically left uninitialized.

This commit also changes the order of the conversion to network byte
order when using the helper function. The footer is now converted before
the header. This eases debugging because the header is still readable
when the footer is converted.