trie: fix deletion of non leaf nodes
When a non leaf node is deleted one if its childs will takes its place.
The selected child must be the one with the shorter prefix length.
If the one with the longer prefix length is selected the node with the
shorter prefix length will be buried, meaning the search algorithm will
not find it anymore.