rtr/packets: add missing break in rtr_pdu_check_size
In the case SERIAL_NOTIFY there is no break causing it to fall through.
If we receive a malformed SERIAL_NOTIFY with the len field set to the
length of CACHE_RESPONSE this would lead to false positive of the pdu
size check. Resulting in possible out of bound reads.