rtr: prefix interval mode symbols
The intervals mode symbols do not have a prefix at all right now. This
is problematic because it could easily lead to name collisions when used
in other projects.
Furthermore are the {set,get}_interval_mode functions in the wrong place
they operate on rtr_sockets and should therefore be in rtr.c not in
rtr_mgr.c

While this is a breaking API change I argue that it does not really
matter because the interval mode code has never officially been
released.