Fix Debian packaging.
This fixes several little things which were preventing the existing
Debian packaging from building cleanly with tools like gbp:

* There was a missing comma in one of the dependency lists in
  debian/control.

* debian/copyright was in a format that Lintian didn't like.

* The version numbering and source package format were wrong: the
  RTRlib project is the original source of this code, so this should
  have been a native package.  This patch changes the source package
  format from (implied) "1.0" to "3.0 (native)".

The change in source package format requires a corresponding change to
the package version numbering: native packages don't use the
"debian_revision" field of the package version number.

This patch adds a dummy 0.5.0.0 "release" so that the most recent
version in debian/changelog will be in the correct format and so that
the resulting package version numbers will sort correctly; feel free
to replace the "0.5.0.0" with the real 0.5.1 entry when that's ready.

A simpler alternative would have been just to retroactively whack all
the listed release numbers into the native format by removing the "-1"
suffix from every entry in debian/changelog, but that change could be
construed as rewriting history, so this patch takes the more
conservative approach.

See the "Version numbering" section of the Debian Policy Manual for
more than you ever wanted to know about Debian package versioning.

With these changes, the build works with gbp and passes Lintian.