rtrlib/rtr: fix race condition in rtr_stop
rtr_stop currently has a race condition. It uses a unsynchronized shared
variable. Which value this variable has is therefore undefined when the
thread that rtr_stop tries to kill reads it. The only direct solution is
a explicit memory barrier, but C does not support them until C11, even
there it's optional.

The new approach uses pthread_cancel and carefully selected code regions
where it is safe to exit without freeing resources. The shared variable
is still used but we do not rely on it for timely thread exit anymore.