third-party: move tommyds from spki/hashtable into third party folder and remove version from folder name
tommyds has not be used exclusively by spki for quite some time, it
therefore does not make sense to keep in under the spki directory.

Furthermore having the tommyds version in the folder name only makes
updating the lib more verbose and is not necessary for knowing the
current version since the version number is available in several places
within files of the library.