rtrlib: support for vrf
vrf support helps user to create sessions on some vrfs.
a new lrtr_vrf api is available, and is extended enough to do:
-1 support for vrf with netns backend
-2 support for vrf lite
-3 api to give permission rights to the library

1 is implemented. While vrf is not available, the tcp/ssh connection is
not created. when vrf is created, then tcp/ssh connection will be tried.
retries are rescheduled as if there was a standard network error ( 600
seconds from bgp rpki side).

2 is implemented too, but not tested.
difficulty resides in that a bind operation must be done after socket
creation, and before socket connection. in ssh , the two operations are
opaque from rtrlib point of view, since it is done in sshlib library.
the solution chosen consists in creating the socket and passing the
socket id as parameter to ssh.

3 is implemented with linux capabilities or user capabilities.
when capabilities are not used in a system, the calling system should
have the user rights to change of netns(). in some daemons, it is usual
to start as root, then change the owner of the daemon to have less
privileges. Then, when some operations need some privileges, a quick
change of owner or privileges is operated. rtrlib has the vrf init API
that is expanded to provide 2 user ids, one which is the initial userid
with all the supposed rights, and the second which is the daemonized
user. Like that, when rtrlib will want to do some privileged operations
to change netns, then seteuid() will be called appropriately.

The change of permission rights is valid while the socket creation is
done in the same pthread. for instance, for ssh session, if the ssh call
to create a connection in libssh creates a new pthread, then the
privileges may be lost, and socket operation mail fail.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>