*: fix warnings of compilation with gcc 5.4
some warnings are removed by adding some tiny modifications in the code.
third party files are omitted in the warnings fix.
also there is a warning in travis when braces are put around fgets(){}
call, whereas the warning had disappeared with gcc 5.4. decision is to
keep gcc warning for this case, so that travis does not warn.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>