rtrlib: ability to pass configured socket in tcp/ssh attribute
Thanks to previous commit, it becomes possible by the caller to add
an an optional opaque pointer, as well as a callback function that
will be in charge of providing a file descriptor that will be used to
create the tcp/ssh connection to the server. If the information is
provided, and if the socket returned is invalid (<=0), then a fallback
mechanism is done, that is to say that the socket is created by rtrlib
itself.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>