rtrlib/transport/{tcp,ssh}: fix tr_{tcp,ssh}_open blocking socket stop
Since the move from pthread_kill to pthread_cancel a rtr_socket could
not be stopped immediately if it was in the tr_open call.
cancelability could not be enabled for that function because it holds
local resources.
This enables cancalability locally for the only blocking call in
tr_tcp_open and tr_ssh_open respectively.