rtrlib/transport/ssh/ssh_transport.c: fix ssh-rsa public key auth
When SSH public key authentication is attempted right after
ssh_connect(), there's a chance `SSH_MSG_EXT_INFO` message will not be
received by the client in timely manner. In case if an `ssh-rsa`
public key is being used, this in turn may lead to libssh picking SHA1
digest algorithm, which is disabled by default:
https://github.com/canonical/libssh/blob/f23d1454e50d0dbb314edd9bf4227ab72303484b/src/pki.c#L443-L454
This may cause intermittent SSH authentication failures with no clear
reason as SSH logging is disabled in rtrlib.
The fix is based on this suggestion:
https://archive.libssh.org/libssh/2025-05/0000023.html